backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
requests==2.28.1
python-dotenv==0.19.0
markdown==3.3.4
